Translate the part of the new features of LXC 2.1 into Japanese
I guess that many Japanese lxc users want to know
about new features of LXC 2.1 as soon as possible,
so I have not finished all the translations of LXC 2.1
announcement yet, but I publish it.

I am going to translate the rest as soon as possible.

This patch also fixes a typo in English announcement.

Reviewed-by: Hiroaki Nakamura <hnakamur@gmail.com>
Signed-off-by: KATOH Yasufumi <karma@jazz.email.ne.jp>