doc: css: fix word wrap in tables
Wrap text in table columns if there is enough space for it,
otherwise the table will be scrollable as before.
The current implementation will wrap if the width is >= 700px
in mobile view or >= 1200px in desktop view.
If the screen is smaller, text would wrap to one word per line,
which isn't readable.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>