remove unnecessary line breaks in Japanese pages
In this change, we have not changed Japanese part. We have removed line
breaks in the commented English section.(Such changes were made in #540)

The Japanese content has not changed, but if the Japanese page is not
updated after English version, "the page is out of date" will be
displayed.

Signed-off-by: KATOH Yasufumi <karma@jazz.email.ne.jp>