Update Vanilla framework to 4.1; change font to Inter
Vanilla 4.1 was released July 31 and doesn't require major changes but
switches to variable font widths. At the same time, I've updated the
site-wide default font to use Inter, a widely-used open source font, to
reduce the Canonical-style branding of the site.

The two other changes I had to make to the static CSS were the max-width
of paragrahs (to not wrap lines with the slightly-wider Inter) and
removing a margin workaround for breadcrumbs as this appears to have
been fixed in Vanilla.

Feedback welcome, especially around linking to the Inter CDN versus
including the font locally—or just switching to system fonts, instead.

Signed-off-by: Cassidy James Blaede <c@ssidyjam.es>