Branding: Update logo
I created this logo from scratch and am licensing it under
CC BY-SA 4.0. It's intended to feel familiar to the previous
Canonical-designed logo while reminding people of Tux the
penguin. :) If you squint hard you might be able to see an L
and C in there, too.

I am not a professional logo designer; I would be more than
happy to let someone else who is more skilled step in. The
original SVG was created in Inkscape, and could be used as a
jumping off point if anyone wants.

Signed-off-by: Cassidy James Blaede <c@ssidyjam.es>